Renamed schemas to types
