+++ src/types/badge.gql
