+++ src/types/friends.gql
