+++ src/types/location.gql
